BasicProtocol (#297)
